[client] | refactor(Profile): 풀 리퀘스트 CL 요구사항 반영
client 에서 풀 리퀘스트의 요구사항을 반영하였습니다.
