+++ packages/client/constants.ts
@@ -92,9 +92,6 @@ export const signUpFormDataMaxLength = {
   introduction: 1500,
 };
 
-export const DEFAULT_USER_IMAGE =
-  'https://icon-library.net/images/small-icon-images/small-icon-images-11.jpg';
-
 export const orientation = {
   vertical: 'vertical',
   horizontal: 'horizontal',
