+++ packages/client/views/Profile/hook/use-profile.ts
@@ -4,49 +4,56 @@ import { useRouter } from 'next/router';
 import { v4 } from 'uuid';
 
 import { endpoint } from '../../../constants';
-import { useUserProfile } from '../hook/use-user-profile';
 import { useUser } from '../../../components/UserProvider/hooks';
 import { makeQueryUserInfoAction } from '../action/make-query-user-info-action';
-import { makeMutateUserAvatarAction } from '../action/make-mutate-user-avatar-action';
-import { makeMutateUserProfileAction } from '../action/make-mutate-user-profile-action';
+import { makeUpdateUserAvatarAction } from '../action/make-update-user-avatar-action';
+import { makeUpdateUserProfileAction } from '../action/make-update-user-profile-action';
 import { makeUploadAvatarToBucketAction } from '../action/make-upload-avatar-to-bucket-action';
-import { makeMutateAvatarPreSignedURLAction } from '../action/make-mutate-avatar-pre-signed-url-action';
+import { makeGetAvatarPreSignedURLAction } from '../action/make-get-avatar-pre-signed-url-action';
+import { UserProfileState } from '../model/user-profile-state';
 
 export const useProfile = () => {
   const user = useUser();
   const router = useRouter();
 
-  const {
-    userProfile,
-    setUserProfile,
-    handleUsername,
-    handleDescription,
-  } = useUserProfile();
+  const [userProfile, setUserProfile] = useState(
+    new UserProfileState(0, '', '', ''),
+  );
 
-  const id = user ? user.id : 16;
-  const queryUserInfoAction = makeQueryUserInfoAction(id);
+  const handleUsername = e => {
+    setUserProfile({
+      ...userProfile,
+      username: e.target.value,
+    });
+  };
+
+  const handleDescription = e => {
+    setUserProfile({
+      ...userProfile,
+      description: e.target.value,
+    });
+  };
+
+  const userId = user ? user.id : 16;
+  const queryUserInfoAction = makeQueryUserInfoAction(userId);
   const { query } = useQuery(queryUserInfoAction, false);
 
   useEffect(() => {
     const initialize = async () => {
-      if (!user) {
-        router.push(endpoint.hotlist);
-      }
+      // if (!user) {
+      //   router.push(endpoint.hotlist);
+      // }
 
-      const { error, payload } = await query();
+      const { error: userUndefinedError, payload } = await query();
 
-      if (error) {
+      if (userUndefinedError) {
+        // Error Control
         return;
       }
 
-      const { username, description, avatar } = payload;
+      const { id, username, description, avatar }: UserProfileState = payload;
 
-      setUserProfile({
-        id,
-        username,
-        description,
-        avatar,
-      });
+      setUserProfile(new UserProfileState(id, avatar, username, description));
     };
 
     initialize();
@@ -54,89 +61,80 @@ export const useProfile = () => {
 
   const [isAvatarFetching, setIsAvatarFetching] = useState(false);
   const [isFormFetching, setIsFormFetching] = useState(false);
-  const [userAvatarUpdateError, setUserAvatarUpdateError] = useState(false);
-  const [userFormUpdateError, setUserFormUpdateError] = useState(false);
 
-  const getAvatarPreSignedUrl = useMutation(makeMutateAvatarPreSignedURLAction)
+  const getAvatarPreSignedUrl = useMutation(makeGetAvatarPreSignedURLAction)
     .mutate;
   const uploadAvatarToBucket = useMutation(makeUploadAvatarToBucketAction)
     .mutate;
-  const changeUserAvatar = useMutation(makeMutateUserAvatarAction).mutate;
-  const changeUserProfile = useMutation(makeMutateUserProfileAction).mutate;
+  const changeUserAvatar = useMutation(makeUpdateUserAvatarAction).mutate;
+  const changeUserProfile = useMutation(makeUpdateUserProfileAction).mutate;
 
-  const handleAvatarSubmit = e => {
+  const handleAvatarSubmit = async e => {
     setIsAvatarFetching(true);
-    if (!e.target.files[0]) {
+
+    const avatarFile = e.target.files[0];
+    if (!avatarFile) {
+      // Error Control
+      return;
+    }
+    const avatarFileName = `${v4()}/${avatarFile.name}`;
+
+    const {
+      payload: avatarPreSignedURL,
+      error: AvatarPreSignedUrlError,
+    } = await getAvatarPreSignedUrl(avatarFileName);
+
+    if (AvatarPreSignedUrlError) {
+      // Error Control
       return;
     }
 
-    const updateAvatar = async () => {
-      const avatarFile = e.target.files[0];
-      const avatarFileName = `${v4()}/${avatarFile.name}`;
+    const { error: AvatarToBucketError } = await uploadAvatarToBucket({
+      preSignedUrl: avatarPreSignedURL,
+      file: avatarFile,
+    });
 
-      const {
-        payload: avatarPreSignedURL,
-        error: getAvatarPreSignedUrlError,
-      } = await getAvatarPreSignedUrl(avatarFileName);
+    if (AvatarToBucketError) {
+      // Error Control
+      return;
+    }
 
-      if (getAvatarPreSignedUrlError) {
-        setUserAvatarUpdateError(true);
-        return;
-      }
+    const url = `${process.env.AVATAR_SOURCE_HOST}/${avatarFileName}`;
 
-      const { error: uploadAvatarToBucketError } = await uploadAvatarToBucket({
-        preSignedUrl: avatarPreSignedURL,
-        file: avatarFile,
-      });
+    const {
+      payload: changedUser,
+      error: changeUserAvatarError,
+    } = await changeUserAvatar({ id: userProfile.id, avatar: url });
 
-      if (uploadAvatarToBucketError) {
-        setUserAvatarUpdateError(true);
-        return;
-      }
+    const isUserAvatarChanged = !changeUserAvatarError && changedUser;
+    if (!isUserAvatarChanged) {
+      // Error Control
+      return;
+    }
 
-      const url = `${process.env.S3_AVATAR_PATH}/${avatarFileName}`;
-
-      const {
-        payload: changedUser,
-        error: changeUserAvatarError,
-      } = await changeUserAvatar({ id: userProfile.id, avatar: url });
-
-      const isUserAvatarChanged = !changeUserAvatarError && changedUser;
-      if (!isUserAvatarChanged) {
-        setUserAvatarUpdateError(true);
-        return;
-      }
-
-      setIsAvatarFetching(false);
-      setUserProfile({
-        ...userProfile,
-        avatar: url,
-      });
-    };
-
-    updateAvatar();
+    setIsAvatarFetching(false);
+    setUserProfile({
+      ...userProfile,
+      avatar: url,
+    });
   };
 
-  const handleFormSubmit = e => {
+  const handleFormSubmit = async e => {
     setIsFormFetching(true);
 
-    const updateUserForm = async () => {
-      const { payload, error } = await changeUserProfile({
-        id: userProfile.id,
-        username: userProfile.username,
-        description: userProfile.description,
-      });
+    const { payload, error } = await changeUserProfile({
+      id: userProfile.id,
+      username: userProfile.username,
+      description: userProfile.description,
+    });
 
-      if (error || !payload) {
-        setUserFormUpdateError(true);
-        setIsFormFetching(false);
-        return;
-      }
-
+    if (error || !payload) {
+      // Error Control
       setIsFormFetching(false);
-    };
+      return;
+    }
 
-    updateUserForm();
+    setIsFormFetching(false);
   };
 
   return {
@@ -146,8 +144,6 @@ export const useProfile = () => {
     handleDescription,
     handleFormSubmit,
     isAvatarFetching,
-    userAvatarUpdateError,
-    userFormUpdateError,
     isFormFetching,
   };
 };
